refactor: 종료미션 페이지 분리
