+++ src/app/result/FinishedMissionList.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function FinishedMissionList() {
+  return <div>FinishedMissionList</div>;
+}
+
+export default FinishedMissionList;
